Move header-only c10/util Layer 1 utils to torch/headeronly
Migrate three already-header-only c10/util headers that depend on headers
migrated earlier in this wave into torch/headeronly/util/, leaving the old c10
paths as one-line forwarders (fully backward compatible, no consumer include
changes):

  static_tracepoint.h, safe_numerics.h, sparse_bitset.h

static_tracepoint.h is a plain move (preprocessor macros only) with its include
of static_tracepoint_elfx86.h repointed at torch/headeronly. safe_numerics is
flipped into torch::headeronly (re-aliased into c10), with its llvmMathExtras
include repointed and its enable_if SFINAE annotated NOLINT(modernize-use-
constraints), matching the existing convention. sparse_bitset is flipped too and
its bare llvm:: references are qualified as c10::llvm:: now that the class lives
in torch::headeronly; as LLVM-derived vendored code it is excluded from
clang-tidy at the new path (mirroring the c10 vendored exclusions).

Deferred from Layer 1 (not cleanly header-only as-is):
- c10/util/generic_math.h includes c10/cuda/CUDAMathCompat.h (and the hip
  equivalent) under __CUDA_ARCH__, so the headeronly file would reference c10
  device headers.
- c10/util/TypeIndex.h parses __PRETTY_FUNCTION__ and hardcodes c10:: type names
  (including c10::basic_string_view and its own c10::util::detail qualified
  name); moving it changes the compiler-printed names and breaks its
  static_asserts unless those strings are carefully rewritten.

Test Plan:

Full build (confirms all libtorch consumers still compile):

```
pip install -e . --no-build-isolation
```

Header-only ABI check (compiles + runs without linking libtorch):

```
build/bin/test_aoti_abi_check --gtest_filter='TestSafeNumerics*:TestSparseBitset*'
```

Lint:

```
lintrunner -a <changed files>
```

This PR was authored with the assistance of Claude (Claude Code).

[ghstack-poisoned]
